+++ packages/stylelint-config/src/bases/base.ts
