+++ packages/stylelint-config/src/bases/extend.ts
