+++ packages/stylelint-config/src/index.ts
@@ -1,3 +1,22 @@
-export { default as astro } from "./astro"
-export { default as css } from "./css"
-export { default as scss } from "./scss"
+import type { Config } from "stylelint"
+
+import { baseConfig } from "./bases/base"
+import { cssExtend } from "./bases/extend"
+
+const { rules: baseRules, ...base } = baseConfig
+
+const css: Config = {
+  ...base,
+  extends: cssExtend,
+  rules: {
+    ...baseRules,
+    "at-rule-no-unknown": [
+      true,
+      {
+        ignoreAtRules: ["tailwind"],
+      },
+    ],
+  },
+}
+
+export default css
